service item listing
