--- src/connection/connection.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ConnectionController } from './connection.controller';
-
-describe('ConnectionController', () => {
-  let controller: ConnectionController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ConnectionController],
-    }).compile();
-
-    controller = module.get<ConnectionController>(ConnectionController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
