--- src/item/item.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ItemController } from './item.controller';
-
-describe('ItemController', () => {
-  let controller: ItemController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ItemController],
-    }).compile();
-
-    controller = module.get<ItemController>(ItemController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
