+++ src/item/item.controller.ts
@@ -5,36 +5,9 @@ import { UserToken } from 'src/auth/dto/usertoken.dto';
 import { ItemService } from './item.service';
 import { response } from 'express';
 import { FilterItemRequestDTO } from './dto/filter-item.dto';
+import { ServiceItemService } from './service-item.service';
 
 @Controller('item')
 export class ItemController {
-    constructor(private serviceItemService: ItemService){
-       
-    }
-
-    @UseGuards(JwtAuthGuard)
-    @Get('serviceItem')
-    async getServiceItems(
-      @Req() req,
-      @Query(ValidationPipe) query: FilterItemRequestDTO,
-      @Query("skip", ParseIntPipe) skip: number,
-      @Query("limit", ParseIntPipe) limit: number,
-      @GetToken() token: UserToken,
-      @Res() res: Response
-    ) {
-      try {
-        let data = await this.serviceItemService.getServiceItems(
-          query,
-          token,
-          req,
-          skip,
-          limit
-        );
-        return data;
-      } catch (err) {
-       
-        
-        return err;
-      }
-    }
+    
 }
