--- src/profile/enum/profile.enum.ts
@@ -1,9 +1,0 @@
-export enum EProfileVisibility {
-  public = "Public",
-  private = "Private",
-}
-
-export const ESProfileVisibility = [
-  EProfileVisibility.public,
-  EProfileVisibility.private,
-];
