--- src/project/interface/project-media-type.enum.ts
@@ -1,4 +1,0 @@
-export enum EProjectMediaType {
-  // TODO: Need to change the value to video project.
-  VideoDocument = "ProjectContent",
-}
