--- src/user/dto/create-peer-tube-user.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsOptional, IsString } from "class-validator";
-
-export class CreatePeerTubeUserDTO {
-  @IsNotEmpty()
-  @IsString()
-  phoneNumber: string;
-
-  @IsOptional()
-  @IsString()
-  emailId?: string;
-}
