--- src/user/enum/number-type.enum.ts
@@ -1,9 +1,0 @@
-export enum EPhoneNumberType {
-  primary = "Primary",
-  secondary = "Secondary",
-}
-
-export const ESPhoneNumberType = [
-  EPhoneNumberType.primary,
-  EPhoneNumberType.secondary,
-];
