--- src/user/enum/user-reference.enum.ts
@@ -1,13 +1,0 @@
-export enum EUserReference {
-  vendor = "vendor",
-  user = "user",
-  customer = "customer",
-  employee = "employee",
-}
-
-export const ESUserReference = [
-  EUserReference.vendor,
-  EUserReference.user,
-  EUserReference.employee,
-  EUserReference.customer,
-];
